directory structure refactoring
